use ParseInt instead of Atoi for parsing durations
Flaeg is failing to deserialize durations with values >= to 2^32 on 32 bits architectures.
This is caused by the usage of `Atoi` to parse duration, which returns an `int` not an `int64`.

During a JSON unmarshall of a duration for values >= to 2^32, Atoi will
return an error and flaeg will try to parse the value as a string, which
is going to fail too, causing the unmarshalling to fail.

Using `ParseInt(s, 10, 64)` fixes the problem, as it works on an int64.

It is even a little bit faster according to [this](https://gist.github.com/evalphobia/caee1602969a640a4530) bench.